fix: some PR comments
